+++ spec/rubocop/cop/lint/useless_access_modifier_spec.rb
@@ -34,6 +34,9 @@ describe Rubocop::Cop::Lint::UselessAccessModifier do
     let(:source) do
       [
         'class SomeClass',
+        '  def some_method',
+        '    puts 10',
+        '  end',
         '  protected',
         'end'
       ]
@@ -44,7 +47,7 @@ describe Rubocop::Cop::Lint::UselessAccessModifier do
       expect(cop.offenses.size).to eq(1)
       expect(cop.offenses.first.message)
         .to eq('Useless `protected` access modifier.')
-      expect(cop.offenses.first.line).to eq(2)
+      expect(cop.offenses.first.line).to eq(5)
       expect(cop.highlights).to eq(['protected'])
     end
   end
@@ -168,6 +171,28 @@ describe Rubocop::Cop::Lint::UselessAccessModifier do
         .to eq('Useless `private` access modifier.')
       expect(cop.offenses.first.line).to eq(2)
       expect(cop.highlights).to eq(['private'])
+    end
+  end
+
+  context 'block modifiers' do
+    let(:source) do
+      [
+        'class SomeClass',
+        '  private',
+        'end'
+      ]
+    end
+
+    let(:corrected_source) do
+      [
+        'class SomeClass',
+        '  ',
+        'end'
+      ].join("\n")
+    end
+
+    it 'auto-corrects' do
+      expect(autocorrect_source(cop, source)).to eq(corrected_source)
     end
   end